requirements.txt
Django 
django-browser-reload 
python-dotenv
python-decouple
django-tinymce
autopep8
pytest 
pytest-django
gunicorn
